requirements.txt
discord.py==1.3.3
python-dotenv
pep8
asyncpg
autopep8
Pillow==7.0.0
git+https://github.com/Rapptz/discord-ext-menus